+++ src/components/postCard/PostMessage.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="border-t flex rounded-b-lg">
+    <input
+      type="text"
+      class="w-full p-4 rounded-b-lg focus:outline-none focus:border-purple-400 border-gray-300"
+      placeholder="Comment This"
+    />
+    <button
+      class="h-fit my-auto text-gray-400 hover:text-gray-600 hover:bg-gray-100 rounded-full px-3 border-2 border-gray-300 mx-3"
+    >
+      Post
+    </button>
+  </div>
+</template>