+++ src/stores/profile.ts
@@ -3,16 +3,19 @@ import type { User } from '@/types/user'
 import { mockProfileData } from '@/utils/mockProfile'
 import { mockNotifications } from '@/utils/mockNotifications'
 import type { Notification } from '@/types/notification'
+import { mockSuggestedPeople } from '@/utils/suggestedPeople'
 
 type ProfileState = {
   user: User
   notifications: Notification[]
+  suggestedPeople: User[]
 }
 
 export const useProfileStore = defineStore('profile', {
   state: (): ProfileState => ({
     user: {} as User,
-    notifications: []
+    notifications: [],
+    suggestedPeople: []
   }),
 
   getters: {
@@ -21,6 +24,9 @@ export const useProfileStore = defineStore('profile', {
     },
     getNotifications(): Notification[] {
       return this.notifications
+    },
+    getSuggestedPeople(): User[] {
+      return this.suggestedPeople
     }
   },
 
@@ -40,6 +46,14 @@ export const useProfileStore = defineStore('profile', {
       this.notifications = mockNotifications
 
       return this.notifications
+    },
+    async fetchSuggestedPeople() {
+      // Simulate delay
+      await new Promise((resolve) => setTimeout(resolve, 1000))
+      // Return the mock user data
+      this.suggestedPeople = mockSuggestedPeople
+
+      return this.suggestedPeople
     }
   }
 })