+++ src/types/user.ts
@@ -1,5 +1,7 @@
 export type User = {
   name: string
-  title: string
+  profession?: string
+  is_doctor?: boolean
+  location?: string
   picture: string
 }