+++ CHANGELOG.md
@@ -1,7 +1,35 @@
 # Changelog
 All significant updates to this project will be meticulously documented in this log.
 
-## 0.0.0.alpha.1 - 19-03-2027
+## 0.1.0 - 20-03-2024
+
+### Added
+
+- **Posts**: Posts have been added to the website.
+
+- **Post Store**: A new store has been added to the project to fetch the posts.
+
+- **Post Skeleton**: A new skeleton for the posts has been added to the website.
+
+- **Swiper**: A new swiper has been added to the website. `swiper/vue`
+
+- **Posts Type**: A new type has been added to the project to represent the posts.
+
+- **Suggested People**: A new section inside posts has been added to suggest people to follow.
+
+### Changed
+
+- **Navigation bar**: The navigation bar has been updated to stick to the top of the page when scrolling.
+
+- **Title field**: The title field of the user type has been renamed to `profession`.
+
+- **Extended User Type**: The user type has been extended to include the `profession`, `is_doctor`, and `location` fields.
+
+### Docs
+
+- **Badge**: New badges has been added to the README file.
+
+## 0.0.0.alpha.1 - 19-03-2024
 
 ### Added